Loading Image Metatags for Profile pages
Refactored the profile pages to prefetch profileData
in order to facilitate correctly blocking page loads
until the correct image metatags have been placed.

Minor bugfixes along the way.
