+++ front-end/src/app/games/hashi/hashi.component.ts
@@ -12,7 +12,7 @@ import { SettingsService } from '../../services/persistence/settings.service';
 import { GameStarterService } from '../../services/generators/game-starter.service';
 import { GameBoard } from '../../classes/game-board';
 import { OptionsService } from '../../services/games/options.service';
-import { MetaService } from '../../services/meta.service';
+import { MetaService } from '../../services/meta/meta.service';
 
 @Component({
   selector: 'app-hashi',
