+++ front-end/src/app/main-menu/main-menu.component.ts
@@ -9,7 +9,7 @@ import { isPlatformBrowser } from '@angular/common';
 import { GameListAllService } from '../services/games/game-list-all.service';
 import { Subscription } from 'rxjs/Subscription';
 import { RelayTrackerService } from '../services/relay/relay-tracker.service';
-import { MetaService } from '../services/meta.service';
+import { MetaService } from '../services/meta/meta.service';
 
 @Component({
   selector: 'app-main-menu',
