+++ front-end/src/app/services/meta/meta.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { MetaService } from './meta.service';
+
+describe('MetaService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: MetaService = TestBed.get(MetaService);
+    expect(service).toBeTruthy();
+  });
+});
