+++ front-end/src/app/services/profile-resolver/profile-resolver.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { ProfileResolverService } from './profile-resolver.service';
+
+describe('ProfileResolverService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: ProfileResolverService = TestBed.get(ProfileResolverService);
+    expect(service).toBeTruthy();
+  });
+});
