+++ front-end/src/index.html
@@ -13,19 +13,20 @@
   <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
 
   <meta name="description" content="">
-  <meta name="og:description" content="">
-  <meta name="twitter:description" content="">
-  <meta name="og:title" content="">
-  <meta name="twitter:title" content="">
-  <meta name="og:type" content="">
-  <meta name="twitter:type" content="">
+  <meta property="og:description" content="">
+  <meta property="twitter:description" content="">
+  <meta property="og:title" content="">
+  <meta property="twitter:title" content="">
+  <meta property="og:type" content="">
+  <meta property="twitter:type" content="">
   <meta name="url" content="">
-  <meta name="og:url" content="">
-  <meta name="twitter:url" content="">
-  <meta name="og:image" content="">
-  <meta name="twitter:image" content="">
-  <meta name="og:image:width" content="">
-  <meta name="og:image:height" content="">
+  <meta property="og:url" content="">
+  <meta property="twitter:url" content="">
+  <meta property="og:image" content="">
+  <meta property="twitter:image" content="">
+  <meta property="og:image:width" content="">
+  <meta property="og:image:height" content="">
+  <meta property="og:image:type" content="" >
   <meta name="keywords" content="">
   <meta name="robots" content="">
 </head>
